+++ src/Entity/Author.php
@@ -22,7 +22,7 @@ class Author
 
     #[ORM\Column(length: 255, unique: true)]
     #[Assert\NotBlank]
-    private ?string $name = null;
+    private string $name = '';
 
     /**
      * @var Collection<int, ComicSeries>
@@ -37,7 +37,7 @@ class Author
 
     public function __toString(): string
     {
-        return $this->name ?? '';
+        return $this->name;
     }
 
     public function getId(): ?int
@@ -45,7 +45,7 @@ class Author
         return $this->id;
     }
 
-    public function getName(): ?string
+    public function getName(): string
     {
         return $this->name;
     }